requirements.txt
Markdown>=3.3.0
PyYAML>=6.0